优化代码

+++ app/src/main/java/com/guyuan/dear/focus/client/data/FocusClientRepository.java
@@ -39,7 +39,7 @@ public class FocusClientRepository {
         return apiService.getClientList(body);
     }
 
-    Observable<ResultBean<ClientCompanyBean>> getClientBasicInfo(int id) {
+    Observable<ResultBean<ClientCompanyBean>> getClientBasicInfo(long id) {
         return apiService.getClientBasicInfo(id);
     }
