+++ app/src/main/java/com/guyuan/dear/work/client/data/WorkClientViewModel.java
@@ -97,7 +97,7 @@ public class WorkClientViewModel extends BaseViewModel {
      *
      * @param id 客户id
      */
-    public void getClientBasicInfo(int id) {
+    public void getClientBasicInfo(long id) {
 
         Disposable disposable = RxJavaHelper.build(this, repository.getClientBasicInfo(id))
                 .getHelper().flow(clientBasicEvent);
